+++ src/main/java/roomescape/repository/implementation/JdbcThemeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.repository.implementation;
 
 import java.util.List;
 import javax.sql.DataSource;
@@ -9,6 +9,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Theme.Theme;
+import roomescape.repository.ThemeRepository;
 
 @Repository
 public class JdbcThemeRepository implements ThemeRepository {
